+++ examples/samples/resources/meraki_networks_group_policies copy/resource.tf
@@ -13,6 +13,6 @@ provider "meraki" {
 }
 
 resource "meraki_networks_group_policies" "foobar" {
-  network_id = "L_828099381482775376"
+  network_id = "L_828099381482775779"
   name = "foobar"
 }