+++ examples/samples/resources/meraki_networks_wireless_billing/resource.tf
@@ -23,9 +23,7 @@ resource "meraki_networks_wireless_billing" "example" {
   currency   = "USD"
   network_id = var.my_network_id
   plans = [{
-
     bandwidth_limits = {
-
       limit_down = 1000
       limit_up   = 1000
     }
@@ -35,7 +33,6 @@ resource "meraki_networks_wireless_billing" "example" {
     {
 
       bandwidth_limits = {
-
         limit_down = 500
         limit_up   = 1000
       }